dash-sync/requirements.txt
streamlit>=1.41
Authlib>=1.3.2
plotly>=5.20
pandas>=2.0
gspread>=6.0
google-auth>=2.20
google-auth-oauthlib>=1.0
